pagination is working in all pages

+++ app/controllers/users_controller.rb
@@ -2,8 +2,8 @@ class UsersController < ApplicationController
   before_filter :authenticate_user!
 
   def index
-    #authorize! :index, @user, :message => 'Not authorized as an administrator.'
-    @users = User.all
+    authorize! :index, @user, :message => 'Not authorized as an administrator.'
+    @users = User.paginate :per_page => 5, :page => params[:page]
   end
 
   def show    
@@ -11,17 +11,17 @@ class UsersController < ApplicationController
   end
 
   def new
-    #authorize! :index, @user, :message => 'Not authorized as an administrator.'
+    authorize! :new, @user, :message => 'Not authorized as an administrator.'
     @user = User.new    
   end
 
   def edit
-    authorize! :index, @user, :message => 'Not authorized as an administrator.'
+    authorize! :edit, @user, :message => 'Not authorized as an administrator.'
     @user = User.find(params[:id])
   end
   
   def update
-    #authorize! :update, @user, :message => 'Not authorized as an administrator.'
+    authorize! :update, @user, :message => 'Not authorized as an administrator.'
     @user = User.find(params[:id])
     if @user.update_attributes(params[:user], :as => :admin)
       redirect_to users_path, :notice => "User updated."
@@ -31,15 +31,13 @@ class UsersController < ApplicationController
   end
     
   def destroy
-    #authorize! :destroy, @user, :message => 'Not authorized as an administrator.'
-    user = User.find(params[:id])
-    user.destroy
-    redirect_to users_path, :notice => "User deleted."
-    #unless user == current_user
-     # user.destroy
-      #redirect_to users_path, :notice => "User deleted."
-    #else
-     # redirect_to users_path, :notice => "Can't delete yourself."
-   # end
+    authorize! :destroy, @user, :message => 'Not authorized as an administrator.'
+    @user = User.find(params[:id])
+    unless user == current_user
+      user.destroy
+      redirect_to users_path, :notice => "User deleted."
+    else
+      redirect_to users_path, :notice => "Can't delete yourself."
+    end
   end
 end
